+++ src/app/admin-dashboard/page.tsx
@@ -1,0 +1,5 @@
+import AdminDashboardForm from "@/components/forms/AdminDashboardForm";
+
+export default function AdminDashboardPage() {
+  return <AdminDashboardForm />;
+}
