+++ src/app/itemManagement/page.tsx
@@ -19,7 +19,6 @@ export default function ItemManagementPage() {
         setSidebarOpen={setSidebarOpen}
         activeMenu={activeMenu}
         setActiveMenu={setActiveMenu}
-        handleLogout={handleLogout}
       />
       <main
         className={`flex-1 p-8 bg-gray-100 min-h-screen relative ${
