+++ tailwind.config.js
@@ -15,6 +15,7 @@ module.exports = {
         success: '#48BB78',
         gradientStart: '#C3B1E1',
         gradientEnd: '#6B46C1',
+        adminDashboar: '#ade8f4'
       },
     },
   },
